app/requirements.txt
# Additional requirements for AgentMove Web Demo

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# HTTP and async
httpx>=0.27.0
python-multipart>=0.0.6

# Type checking
pydantic>=2.5.0

# Environment variables
python-dotenv>=1.0.0
